azure-fastapi-function/requirements.txt
fastapi
requests
azure-functions
mangum
uvicorn
python-multipart
